// frontend/src/app/api/model/abstractAccount.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AbstractAccount { 
    /**
     * Whether this account is active or not
     */
    actif?: boolean;
    /**
     * Whether this account should be displayed before others
     */
    pinned?: boolean;
    /**
     * Whether this account depends on MiNET\'s main account
     */
    compteCourant?: boolean;
    /**
     * The type of this account
     */
    accountType?: number;
    /**
     * The date-time at which this account was first created
     */
    creationDate?: string;
    /**
     * The member this account belongs to
     */
    member?: number | null;
    /**
     * The name of this account
     */
    name?: string;
    /**
     * The unique identifier of this account
     */
    readonly id?: number;
    /**
     * The current balance of this account
     */
    readonly balance?: number | null;
    /**
     * The pending balance of this account
     */
    readonly pendingBalance?: number | null;
}